# vulkan: support mixed/deepseekR1 FA head sizes