@ubergarm 2025-07-02
I just tried this PR out on my rig with 3090TI FE 24GB VRAM @ 450 Watts. I believe the model I tested has the same base architecture as DeepSeek-V2-Lite. (didn't have a moment to find the exact same model or quantize my own):

* general.architecture str           = deepseek2
* print_info: n_embd_head_k    = 192
* print_info: n_embd_head_v    = 128

I tested tip of main and this PR (with llama-sweep-bench rebased on top). Not sure why CUDA was so bad, I hadn't run this old model in a long time.

![sweep-bench-jeffbolznv-pr14509](https://github.com/user-attachments/assets/6b05f49b-8958-4814-aa1f-7b919c5ad702)

Only tested `FA=1` for all cases.

<details>

<summary>👈 commands and raw data</summary>

```bash
model=/models/bartowski/DeepSeek-Coder-V2-Lite-Instruct-GGUF/DeepSeek-Coder-V2-Lite-Instruct-IQ4_XS.gguf

./build/bin/llama-sweep-bench \
  --model "$model" \
  -fa \
  -c 8704 \
  -ngl 99 \
  --threads 1
```

## llama.cpp:master@e75ba4c04 CUDA 12.9, Driver 575.64
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    0.929 |   550.86 |    1.384 |    92.46 |
|   512 |    128 |    512 |    2.538 |   201.76 |    2.649 |    48.32 |
|   512 |    128 |   1024 |    4.322 |   118.46 |    3.945 |    32.45 |
|   512 |    128 |   1536 |    6.122 |    83.63 |    5.196 |    24.63 |
|   512 |    128 |   2048 |    7.879 |    64.98 |    6.365 |    20.11 |
|   512 |    128 |   2560 |    9.575 |    53.47 |    7.745 |    16.53 |
|   512 |    128 |   3072 |   11.207 |    45.68 |    9.081 |    14.09 |
|   512 |    128 |   3584 |   12.794 |    40.02 |   10.341 |    12.38 |
|   512 |    128 |   4096 |   14.499 |    35.31 |   11.422 |    11.21 |
|   512 |    128 |   4608 |   16.308 |    31.39 |   12.813 |     9.99 |
|   512 |    128 |   5120 |   17.952 |    28.52 |   14.170 |     9.03 |
|   512 |    128 |   5632 |   19.612 |    26.11 |   15.400 |     8.31 |
|   512 |    128 |   6144 |   21.541 |    23.77 |   17.096 |     7.49 |
|   512 |    128 |   6656 |   23.076 |    22.19 |   18.246 |     7.02 |
|   512 |    128 |   7168 |   24.773 |    20.67 |   19.258 |     6.65 |
|   512 |    128 |   7680 |   26.433 |    19.37 |   20.275 |     6.31 |
|   512 |    128 |   8192 |   28.121 |    18.21 |   21.503 |     5.95 |

## llama.cpp:master@e75ba4c04 ggml_vulkan: 0 = NVIDIA GeForce RTX 3090 Ti (NVIDIA) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 49152 | int dot: 1 | matrix cores: NV_coopmat2
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    0.922 |   555.46 |    1.569 |    81.57 |
|   512 |    128 |    512 |    2.351 |   217.78 |    2.962 |    43.21 |
|   512 |    128 |   1024 |    4.126 |   124.09 |    4.375 |    29.26 |
|   512 |    128 |   1536 |    6.162 |    83.09 |    5.793 |    22.09 |
|   512 |    128 |   2048 |    8.066 |    63.48 |    7.096 |    18.04 |
|   512 |    128 |   2560 |    9.584 |    53.42 |    8.441 |    15.16 |
|   512 |    128 |   3072 |   11.095 |    46.15 |    9.788 |    13.08 |
|   512 |    128 |   3584 |   12.726 |    40.23 |   11.117 |    11.51 |
|   512 |    128 |   4096 |   14.336 |    35.71 |   12.422 |    10.30 |
|   512 |    128 |   4608 |   15.930 |    32.14 |   13.828 |     9.26 |
|   512 |    128 |   5120 |   17.542 |    29.19 |   15.256 |     8.39 |
|   512 |    128 |   5632 |   18.875 |    27.13 |   16.631 |     7.70 |
|   512 |    128 |   6144 |   20.451 |    25.04 |   18.109 |     7.07 |
|   512 |    128 |   6656 |   22.306 |    22.95 |   19.287 |     6.64 |
|   512 |    128 |   7168 |   23.845 |    21.47 |   20.734 |     6.17 |
|   512 |    128 |   7680 |   25.203 |    20.32 |   22.073 |     5.80 |
|   512 |    128 |   8192 |   26.900 |    19.03 |   23.636 |     5.42 |

## jeffbolznv/llama.cpp:fa_head_sizes@7912c9f8 ggml_vulkan: 0 = NVIDIA GeForce RTX 3090 Ti (NVIDIA) | uma: 0 | fp16: 1 | warp size: 32 | shared memory: 49152 | int dot: 1 | matrix cores: NV_coopmat2
|    PP |     TG |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |
|-------|--------|--------|----------|----------|----------|----------|
|   512 |    128 |      0 |    0.194 |  2640.74 |    0.838 |   152.83 |
|   512 |    128 |    512 |    0.196 |  2618.82 |    0.825 |   155.20 |
|   512 |    128 |   1024 |    0.198 |  2591.33 |    0.847 |   151.19 |
|   512 |    128 |   1536 |    0.200 |  2564.36 |    0.879 |   145.68 |
|   512 |    128 |   2048 |    0.203 |  2524.33 |    0.905 |   141.37 |
|   512 |    128 |   2560 |    0.207 |  2478.36 |    0.936 |   136.72 |
|   512 |    128 |   3072 |    0.209 |  2452.19 |    0.972 |   131.62 |
|   512 |    128 |   3584 |    0.208 |  2466.19 |    0.982 |   130.31 |
|   512 |    128 |   4096 |    0.211 |  2422.32 |    1.021 |   125.39 |
|   512 |    128 |   4608 |    0.213 |  2401.86 |    1.041 |   122.95 |
|   512 |    128 |   5120 |    0.213 |  2402.19 |    1.072 |   119.40 |
|   512 |    128 |   5632 |    0.215 |  2383.17 |    1.099 |   116.50 |
|   512 |    128 |   6144 |    0.217 |  2358.98 |    1.103 |   116.09 |
|   512 |    128 |   6656 |    0.219 |  2337.41 |    1.133 |   112.94 |
|   512 |    128 |   7168 |    0.221 |  2313.04 |    1.163 |   110.07 |
|   512 |    128 |   7680 |    0.224 |  2285.53 |    1.184 |   108.10 |
|   512 |    128 |   8192 |    0.225 |  2278.68 |    1.213 |   105.51 |

</details>